+++ Source/MQTTnet/Client/Options/MqttClientTlsOptions.cs
@@ -14,6 +14,8 @@ namespace MQTTnet.Client
     {
         public Func<MqttClientCertificateValidationEventArgs, bool> CertificateValidationHandler { get; set; }
 
+        public Func<MqttClientCertificateSelectionEventArgs, X509Certificate> CertificateSelectionHandler { get; set; }
+
         public bool UseTls { get; set; }
 
         public bool IgnoreCertificateRevocationErrors { get; set; }